multi: move offers encoding and signature validation to lnwire
