+++ lnwire/offer.go
@@ -1,11 +1,9 @@
-package offers
+package lnwire
 
 import (
 	"bytes"
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"io"
 	"time"
 
 	"github.com/btcsuite/btcd/btcec/v2"
@@ -48,16 +46,6 @@ const (
 )
 
 var (
-	// lightningTag is the top level tag used to tag signatures on offers.
-	lightningTag = []byte("lightning")
-
-	// offerTag is the message tag used to tag signatures on offers.
-	offerTag = []byte("offer")
-
-	// signatureTag is the field tag used to tag signatures (TLV type= 240)
-	// for offers.
-	signatureTag = []byte("signature")
-
 	// ErrNodeIDRequired is returned when a node pubkey is not provided
 	// for an offer. Note that when blinded paths are supported, we can
 	// relax this requirement.
@@ -74,14 +62,6 @@ var (
 	// ErrInvalidOfferSig is returned when the signature for an offer is
 	// invalid for the merkle root we have calculated.
 	ErrInvalidOfferSig = errors.New("invalid offer signature")
-
-	// ErrInvalidOfferStr is returned when we fail to decode a bech32
-	// encoded offer string.
-	ErrInvalidOfferStr = errors.New("invalid offer string")
-
-	// ErrBadHRP is returned when an offer string has the wrong bech32
-	// human readable prefix.
-	ErrBadHRP = fmt.Errorf("incorrect bech32 hrp, should be: %v", offerHRP)
 )
 
 // Offer represents a bolt 12 offer.
@@ -206,25 +186,6 @@ func (o *Offer) records() ([]tlv.Record, error) {
 	return records, nil
 }
 
-// signatureDigest returns the tagged merkle root that is used for offer
-// signatures.
-func signatureDigest(messageTag, fieldTag []byte,
-	root chainhash.Hash) chainhash.Hash {
-
-	// The tag has the following format:
-	// lightning || message tag || field tag
-	tags := [][]byte{
-		lightningTag, messageTag, fieldTag,
-	}
-
-	// Create a tagged hash with the merkle root.
-	digest := chainhash.TaggedHash(
-		bytes.Join(tags, []byte{}), root[:],
-	)
-
-	return *digest
-}
-
 // Validate performs the validation outlined in the specification for offers.
 func (o *Offer) Validate() error {
 	// At present, we only support offers that contain node IDs because
@@ -267,60 +228,6 @@ func (o *Offer) Validate() error {
 	}
 
 	return nil
-}
-
-func validateSignature(signature [64]byte, nodeID *btcec.PublicKey,
-	digest []byte) error {
-
-	sig, err := schnorr.ParseSignature(signature[:])
-	if err != nil {
-		return fmt.Errorf("invalid signature: %v: %w",
-			hex.EncodeToString(signature[:]), err)
-	}
-
-	if !sig.Verify(digest, nodeID) {
-		return fmt.Errorf("%w: %v for: %v from: %v", ErrInvalidOfferSig,
-			hex.EncodeToString(signature[:]),
-			hex.EncodeToString(digest),
-			hex.EncodeToString(schnorr.SerializePubKey(nodeID)),
-		)
-	}
-
-	return nil
-}
-
-// encodeTU64 encodes a truncated uint64 tlv.
-//
-// Note: lnd doesn't have this functionality on its own yet (only in mpp encode)
-// so it is added here.
-func encodeTU64(w io.Writer, val interface{}, buf *[8]byte) error {
-	if v, ok := val.(*uint64); ok {
-		return tlv.ETUint64T(w, *v, buf)
-	}
-
-	return tlv.NewTypeForEncodingErr(val, "tu64")
-}
-
-// decodeTU64 decodes a truncated uint64 tlv.
-//
-// Note: lnd doesn't have this functionality on its own yet (only in mpp decode)
-// so it is added here.
-func decodeTU64(r io.Reader, val interface{}, buf *[8]byte, l uint64) error {
-	if v, ok := val.(*uint64); ok && 1 <= l && l <= 8 {
-		if err := tlv.DTUint64(r, v, buf, l); err != nil {
-			return err
-		}
-
-		return nil
-	}
-
-	return tlv.NewTypeForDecodingErr(val, "tu64", l, l)
-}
-
-func tu64Record(tlvType tlv.Type, value *uint64) tlv.Record {
-	return tlv.MakeDynamicRecord(tlvType, value, func() uint64 {
-		return tlv.SizeTUint64(*value)
-	}, encodeTU64, decodeTU64)
 }
 
 // EncodeOffer encodes an offer.
