+++ offers/encoding.go
@@ -1,14 +1,26 @@
 package offers
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/btcsuite/btcd/btcutil/bech32"
+	"github.com/carlakc/boltnd/lnwire"
+)
+
+var (
+	// ErrInvalidOfferStr is returned when we fail to decode a bech32
+	// encoded offer string.
+	ErrInvalidOfferStr = errors.New("invalid offer string")
+
+	// ErrBadHRP is returned when an offer string has the wrong bech32
+	// human readable prefix.
+	ErrBadHRP = fmt.Errorf("incorrect bech32 hrp, should be: %v", offerHRP)
 )
 
 // DecodeOfferStr decodes a bech32 encoded offer string, returning our offer
 // type with the information contained in the offer.
-func DecodeOfferStr(offerStr string) (*Offer, error) {
+func DecodeOfferStr(offerStr string) (*lnwire.Offer, error) {
 	// First, strip any joining characters / spare whitespace from the
 	// offer.
 	cleanOffer, err := stripOffer(offerStr)
@@ -30,7 +42,7 @@ func DecodeOfferStr(offerStr string) (*Offer, error) {
 		return nil, fmt.Errorf("convert bits: %w", err)
 	}
 
-	offer, err := DecodeOffer(offerBytes)
+	offer, err := lnwire.DecodeOffer(offerBytes)
 	if err != nil {
 		return nil, fmt.Errorf("could not decode offer: %w", err)
 	}
